+++ ldclient_gocontext.go
@@ -12,9 +12,9 @@ type scopedClientKey struct{}
 //	ctx := ld.GoContextWithScopedClient(context.Background(), scopedClient)
 //	otherFunction(ctx)
 //
-// This function is not stable, and not subject to any backwards compatibility
-// guarantees or semantic versioning. It is not suitable for production usage. Do
-// not use it. You have been warned.
+// This function is in beta. It is still undergoing testing and active
+// development. Its functionality may change without notice, including becoming
+// backwards incompatible.
 func GoContextWithScopedClient(ctx context.Context, client *LDScopedClient) context.Context {
 	return context.WithValue(ctx, scopedClientKey{}, client)
 }
@@ -32,9 +32,9 @@ func GoContextWithScopedClient(ctx context.Context, client *LDScopedClient) cont
 //		}
 //	}
 //
-// This function is not stable, and not subject to any backwards compatibility
-// guarantees or semantic versioning. It is not suitable for production usage. Do
-// not use it. You have been warned.
+// This function is in beta. It is still undergoing testing and active
+// development. Its functionality may change without notice, including becoming
+// backwards incompatible.
 func GetScopedClient(ctx context.Context) (*LDScopedClient, bool) {
 	client, ok := ctx.Value(scopedClientKey{}).(*LDScopedClient)
 	return client, ok
@@ -49,9 +49,9 @@ func GetScopedClient(ctx context.Context) (*LDScopedClient, bool) {
 //		// handle err as appropriate...
 //	}
 //
-// This function is not stable, and not subject to any backwards compatibility
-// guarantees or semantic versioning. It is not suitable for production usage. Do
-// not use it. You have been warned.
+// This function is in beta. It is still undergoing testing and active
+// development. Its functionality may change without notice, including becoming
+// backwards incompatible.
 func MustGetScopedClient(ctx context.Context) *LDScopedClient {
 	client, ok := GetScopedClient(ctx)
 	if !ok {